+++ setup.py
@@ -18,6 +18,7 @@ def main():
         "coqpit",
         "numpy",
         "optuna",
+        "numba <= 0.53.1",
         "opuslib == 2.0.0",
         "pandas",
         "progressbar2",
